pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "summarizer"
version = "0.1.0"
description = "Summarize YouTube videos, local media, or text via OpenAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.0",
    "python-dotenv>=1.0",
    "yt-dlp>=2024.0",
]

[project.optional-dependencies]
local-transcribe = [
    "mlx-whisper>=0.4",
]

[project.scripts]
summarize = "summarizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
